+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # 1.7
 
  - Citizens NPCs are ignored by spawn rules (by default, use target_npc: true to change)
+ - Update for 1.13+
+ - Chunk generator removed (wasn't really ever used)
 
 # 1.6